refactor: rename StoryAceEditor
